[MacDev] Refactor the add/remove buttons used in the plist editor
It is currently a little confusing as to where exactly a new item will
appear when using the + and - buttons to add remove keys. For example
if I click on a key which is an array type and then click +, i'd expect
a new item to be added to the array. Instead, we get a new item added
at the same level as the key I have selected.

To attempt to make this more intuitive we now render a - button on items
which can be removed and we have a special placeholder which is added to
the end of each array/dictionary which is used to add a new item to the
array/dictionary. Fix for #1858.