[MacDev] Implement a helper method for enumerating PObjects
There are several places which need to iterate over a pobject if it
is either a PArray or PDictionary. We typically need both the Identifier
and the PObject itself, so we enumerate as a KeyValuePair. If the object
is a dictionary, we can simply use the Key and Value of the dictionary itself
If the object is a PArray, we can get the Identifier by getting the 'Value'
property for any PValueObjects.