[MacDev] Handle the difference between 'no values left to add' and 'no idea what to add'
If we have no idea what to add (i.e. the key is not part of the schema), then we should
just add a regular PString/whatever and let the user change it as required. If the key
is part of the schema, but we have used all the allowed values, just do nothing.